requirements.txt
absl-py #@ file:///tmp/build/80754af9/absl-py_1607439979954/work
astor==0.8.1
astunparse==1.6.3
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click #@ file:///home/linux1/recipes/ci/click_1610990599742/work
cycler==0.10.0
Flask==1.1.2
flatbuffers==1.12
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-metadata #@ file:///tmp/build/80754af9/importlib-metadata_1602276842396/work
itsdangerous==1.1.0
Jinja2 #@ file:///home/linux1/recipes/ci/jinja2_1610990516718/work
Keras==2.3.1
Keras-Applications #@ file:///tmp/build/80754af9/keras-applications_1594366238411/work
Keras-Preprocessing==1.1.2
kiwisolver #@ file:///C:/ci/kiwisolver_1603996656077/work
Markdown #@ file:///C:/ci/markdown_1605111171606/work
MarkupSafe==1.1.1
matplotlib #@ file:///C:/ci/matplotlib-base_1603357981283/work
mkl-fft==1.0.6
mkl-random==1.0.1.1
#mkl-service==2.3.0
numpy #@ file:///C:/ci/numpy_and_numpy_base_1603480701039/work
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.3.0
pandas #@ file:///C:/ci/pandas_1602088210462/work
Pillow #@ file:///C:/ci/pillow_1609842634117/work
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing #@ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
pyreadline==2.1
python-dateutil==2.8.1
pytz #@ file:///tmp/build/80754af9/pytz_1608922264688/work
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
scipy #@ file:///C:/ci/scipy_1597675683670/work
six #@ file:///C:/ci/six_1605187303045/work
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
tornado #@ file:///C:/ci/tornado_1606942379977/work
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
zipp #@ file:///tmp/build/80754af9/zipp_1604001098328/work
